fix: Correct timer overflow and LOAD_RAM_TIMEOUT_PER_MB value
Cast ms to int64_t before multiplying by 1000 in all ESP32 start_timer()
implementations to prevent uint32_t overflow for timeouts above ~4 seconds.

Fix LOAD_RAM_TIMEOUT_PER_MB from 2000000 to 2000 ms/MB; the previous value
(33 min/MB) was a copy-paste error — 2 s/MB is consistent with the other
per-MB timeout constants.

+++ port/esp32_port.c
@@ -155,7 +155,7 @@ static void esp32_uart_delay_ms(esp_loader_port_t *port, uint32_t ms)
 static void esp32_uart_start_timer(esp_loader_port_t *port, uint32_t ms)
 {
     esp32_port_t *p = container_of(port, esp32_port_t, port);
-    p->_time_end = esp_timer_get_time() + ms * 1000;
+    p->_time_end = esp_timer_get_time() + (int64_t)ms * 1000;
 }
